+++ Plugins/TableProPluginKit/PluginRoutineInfo.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public struct PluginRoutineInfo: Codable, Sendable {
+    public let name: String
+    public let type: String
+
+    public init(name: String, type: String) {
+        self.name = name
+        self.type = type
+    }
+}