+++ TablePro/Models/Schema/RoutineDetailTab.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+enum RoutineDetailTab: String, CaseIterable, Hashable {
+    case definition = "Definition"
+    case parameters = "Parameters"
+    case ddl = "DDL"
+}